Move Echo to another file (WIP); "pass" functions for relayer and routine
